REALTIMEIMAGECLASSIFICATION/requirement.txt
Flask==2.2.2
TensorFlow==2.10.0
Keras==2.10.0
Pillow==9.2.0
NumPy==1.23.4
